Fix problems caused by OAuth requirements.
OAuth is now required for Helix Twitch API Endpoints:

    https://discuss.dev.twitch.tv/t/requiring-oauth-for-helix-twitch-api-endpoints/23916

This broke my code and took a lot of effort to fix. Also,
I had to open this support request which was not as helpful
as I would have liked:

    https://discuss.dev.twitch.tv/t/programattically-get-user-oauth-token-for-subscriber-notifications/26172

I am saving this working code, but will remove some unnecessary
changes in the next commit.